pydxanalyze/requirements.txt
click==7.1.1
cycler==0.10.0
lxml==4.5.0
numpy==1.20.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-pptx==0.6.18
pytz==2019.3
six==1.14.0
XlsxWriter==1.2.8
pandas==1.2.2
matplotlib==3.3.4
